more css, new event form, needs more work on new event

+++ server/app.py
@@ -281,43 +281,53 @@ class FamiliesById(Resource):
             print(f'error occured: {e}')
             return make_response({'error': 'Family not found'}, 404)
         
-    def patch(self, id):
-        try:
-            family = db.session.execute(db.select(Family).filter_by(id=id)).scalar_one()
-            param = request.json
-            for attr in param:
-                setattr(family, attr, param['attr'])
-            db.session.commit()
-            return make_response(family.to_dict(), 202)
-        except NoResultFound:
-            return make_response({'error': 'Family not found'}, 404)
-        except Exception as e:
-            print(f'error occured: {e}')
-            return make_response({'error': ['validation errors']}, 400)
+    # def patch(self, id):
+    #     try:
+    #         family = db.session.execute(db.select(Family).filter_by(id=id)).scalar_one()
+    #         param = request.json
+    #         for attr in param:
+    #             setattr(family, attr, param['attr'])
+    #         db.session.commit()
+    #         return make_response(family.to_dict(), 202)
+    #     except NoResultFound:
+    #         return make_response({'error': 'Family not found'}, 404)
+    #     except Exception as e:
+    #         print(f'error occured: {e}')
+    #         return make_response({'error': ['validation errors']}, 400)
     
     def delete(self, id):
         try:
-            family = db.session.execute(db.select(Family).filter_by(id=id)).scalar_one()
-            db.session.delete(family)
-            db.session.commit()
-            return make_response(jsonify(''), 200)
+            user_id = session.get('user_id')
+            if user_id:
+                family = db.session.execute(db.select(Family).filter_by(id=id)).scalar_one()
+                family_memberships = db.session.execute(db.select(FamilyMember).filter_by(member_type="child", family_id=family.id)).scalars()
+                children = db.session.execute(db.select(Child).select_from(FamilyMember).join(Child, db.and_(Child.id == FamilyMember.member_id, FamilyMember.member_type == "child")).filter(FamilyMember.family_id==family.id)).scalars()
+                db.session.delete(family)
+                for family_membership in family_memberships:
+                    db.session.delete(family_membership)
+                for child in children:
+                    db.session.delete(child)
+                db.session.commit()
+                return make_response(jsonify(''), 200)
+            return make_response({'error': 'No authorization'}, 401)
         except Exception as e:
             print(f'error occured: {e}')
             return make_response({'error': 'Family not found'}, 404)
 
-class FamiliesByUserId(Resource):
-    def get(self, id):
-        try:
-            families = db.session.execute(db.select(Family).filter_by(user_id=id)).scalars()
-            list_family = [fam.to_dict() for fam in families]
-            return make_response(list_family)
-        except Exception as e:
-            print(f'error occured: {e}')
-            return make_response({'error': 'Family not found'}, 404)
+# class FamiliesByUserId(Resource):
+#     def get(self, id):
+#         try:
+#             families = db.session.execute(db.select(Family).filter_by(user_id=id)).scalars()
+#             list_family = [fam.to_dict() for fam in families]
+#             return make_response(list_family)
+#         except Exception as e:
+#             print(f'error occured: {e}')
+#             return make_response({'error': 'Family not found'}, 404)
+# api.add_resource(FamiliesByUserId, '/users-families/<int:id>')
 
 api.add_resource(Families, '/families')
 api.add_resource(FamiliesById, '/families/<int:id>')
-api.add_resource(FamiliesByUserId, '/users-families/<int:id>')
+
 
 class FamilyMembers(Resource):
     def get(self):
@@ -329,15 +339,16 @@ class FamilyMembers(Resource):
             print(f'error occured: {e}')
             return make_response({'error': 'Family not found'}, 404)
 
-class Files(Resource):
-    def get(self):
-        try:
-            files = db.session.execute(db.select(File)).scalars()
-            file = [file.to_dict() for file in files]
-            return make_response(file)
-        except Exception as e:
-            print(f'error occured: {e}')
-            return make_response({'error': 'files not found'}, 404)
+# class Files(Resource):
+#     def get(self):
+#         try:
+#             files = db.session.execute(db.select(File)).scalars()
+#             file = [file.to_dict() for file in files]
+#             return make_response(file)
+#         except Exception as e:
+#             print(f'error occured: {e}')
+#             return make_response({'error': 'files not found'}, 404)
+# api.add_resource(Files, '/child-files')
 
 class FilesByChildId(Resource):
     def get(self, id):
@@ -385,7 +396,6 @@ class FilesByChildId(Resource):
             print(f'error occured: {e}')
             return make_response({'error': 'files not found'}, 404)
         
-api.add_resource(Files, '/child-files')
 api.add_resource(FilesByChildId, '/child-files/<int:id>')
 
 class Events(Resource):
@@ -424,15 +434,7 @@ class Events(Resource):
 
 api.add_resource(Events, '/events')
 
-class EventsById(Resource):
-    # def get(self, id):
-    #     try:
-    #         event = db.session.execute(db.select(Event).filter_by(id=id)).scalar_one()
-    #         return make_response(event.to_dict())
-    #     except Exception as e:
-    #         print(f'error occured: {e}')
-    #         return make_response({'error': 'Event not found'}, 404)
-        
+class EventsById(Resource):      
     def patch(self, id):
         try:
             event = db.session.execute(db.select(Event).filter_by(id=id)).scalar_one()
